projeto_cury_company/requirements.txt
folium==0.20.0
haversine==2.9.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==2.3.5
pandas==2.3.3
pillow==12.1.0
plotly==6.5.0
streamlit==1.52.2
streamlit-folium==0.25.3
